Our save data just graduated from the School of Document Agnosticism. Summa Cum Laude in Flexibility! 🎓🎉
updated and made the save data better with making it document agnostic

+++ source/Agents/GP/CGPAgent.hpp
@@ -81,7 +81,7 @@ namespace cowboys {
     }
 
 
-    void Serialize(tinyxml2::XMLDocument& doc, tinyxml2::XMLElement* parentElem, int fitness = -1) override {
+    void Serialize(tinyxml2::XMLDocument& doc, tinyxml2::XMLElement* parentElem, double fitness = -1) override {
       auto agentElem = doc.NewElement("CGPAgent");
       parentElem->InsertEndChild(agentElem);
